Re-add Conflicts between knm and plasmoid package
As the NM kded module will claim the NetworkManagerUserSettings bus
name and it is auto-loaded by default, it will break knm unless the user
explicitly disables the kded service. We avoid this situation be
re-adding the Conflicts between those two packages until the plasmoid
has been fixed to load the kded service on-demand.